add authorization for delete action

+++ app/Http/Controllers/OfferController.php
@@ -114,6 +114,8 @@ class OfferController extends Controller
      */
     public function destroy(Offer $offer)
     {
+        $this->authorize('delete', $offer);
+
         $offer->delete();
 
         return redirect()->back()
